+++ doc/source/conf.py
@@ -12,6 +12,7 @@
 # serve to show the default.
 
 import sys, os
+import warnings
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -64,9 +65,18 @@ sphinx_gallery_conf = {
     'reference_url': {'matplotlib': 'http://matplotlib.org',
                       'numpy': 'http://docs.scipy.org/doc/numpy',
                       'scipy': 'http://docs.scipy.org/doc/scipy/reference',
+                      'pyproj': 'http://pyproj4.github.io/pyproj/stable/',
                       'geopandas': None},
     'backreferences_dir': 'reference'
 }
+
+# suppress matplotlib warning in examples
+warnings.filterwarnings(
+    "ignore",
+    category=UserWarning,
+    message="Matplotlib is currently using agg, which is a"
+    " non-GUI backend, so cannot show the figure.",
+)
 
 # The suffix of source filenames.
 source_suffix = '.rst'